requirements.txt
matplotlib
numba
numpy
biopython
scipy
dca==1.0.0 -e https://github.com/utdal/py-mfdca@v1.0.0
